fix : Vendor and Shop Repository Interfaces move to Domain Package
