--- order-api/order-infrastructure/src/build.gradle
@@ -1,6 +1,0 @@
-dependencies {
-    implementation project(':model')
-    implementation project(':order-api:order-domain')
-    implementation project(':order-api:order-application')
-    implementation project(':order-api:order-presentation')
-}
