--- order-api/order-presentation/src/build.gradle
@@ -1,4 +1,0 @@
-dependencies {
-    implementation project(':model')
-    implementation project(':order-api:order-application')
-}
