+++ shop-api/src/main/java/org/healthplus/shop/application/service/VendorRetrievalService.java
@@ -2,7 +2,7 @@ package org.healthplus.shop.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.healthplus.shop.domain.exception.VendorNotFoundException;
-import org.healthplus.shop.infrastructure.VendorRepository;
+import org.healthplus.shop.domain.repository.VendorRepository;
 import org.healthplus.shop.domain.entity.Vendor;
 import org.springframework.stereotype.Service;
