+++ shop-api/src/main/java/org/healthplus/shop/domain/repository/ShopRepository.java
@@ -1,4 +1,4 @@
-package org.healthplus.shop.infrastructure;
+package org.healthplus.shop.domain.repository;
 
 import org.healthplus.shop.domain.entity.Menu;
 import org.healthplus.shop.domain.entity.Shop;
@@ -33,4 +33,6 @@ public interface ShopRepository {
   void removeMenu(Menu menu);
 
   List<Menu> findMenus(Long shopId, int start, int size);
+
+  List<Menu> findAllMenus(Long shopId);
 }
