+++ shop-api/src/main/java/org/healthplus/shop/domain/repository/VendorRepository.java
@@ -1,4 +1,4 @@
-package org.healthplus.shop.infrastructure;
+package org.healthplus.shop.domain.repository;
 
 import org.healthplus.shop.domain.entity.Vendor;
 import java.util.Optional;
